Update URL Launcher : Refactored Logic with Koin AndroidContext Injection for Android Module
