+++ composeApp/src/androidMain/kotlin/com/yong/taximeter/common/util/UrlLauncher.android.kt
@@ -3,11 +3,15 @@ package com.yong.taximeter.common.util
 import android.content.Context
 import android.content.Intent
 import androidx.core.net.toUri
+import org.koin.core.component.KoinComponent
+import org.koin.core.component.inject
+import kotlin.getValue
 
 @Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-actual class UrlLauncher(
-    private val context: Context
-) {
+actual object UrlLauncher: KoinComponent {
+    // Android Context
+    private val context: Context by inject()
+
     actual fun openUrl(url: String) {
         val intent = Intent(Intent.ACTION_VIEW, url.toUri()).apply {
             addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
