--- composeApp/src/androidMain/kotlin/com/yong/taximeter/di/PlatformModule.android.kt
@@ -1,9 +1,0 @@
-package com.yong.taximeter.di
-
-import com.yong.taximeter.common.util.UrlLauncher
-import org.koin.core.module.Module
-import org.koin.dsl.module
-
-actual val platformModule: Module = module {
-    single { UrlLauncher(get()) }
-}
