+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/common/def/UrlDef.kt
@@ -1,0 +1,9 @@
+package com.yong.taximeter.common.def
+
+object UrlDef {
+    const val URL_DEVELOPER_BLOG = "https://dev-lr.com"
+    const val URL_DEVELOPER_GITHUB = "https://github.com/yymin1022"
+    const val URL_DEVELOPER_INSTAGRAM = "https://instagram.com/useful_min"
+    const val URL_PRIVACY_POLICY = "https://defcon.or.kr/privacy"
+
+}
