+++ composeApp/src/commonMain/kotlin/com/yong/taximeter/common/util/UrlLauncher.kt
@@ -1,0 +1,6 @@
+package com.yong.taximeter.common.util
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+expect object UrlLauncher {
+    fun openUrl(url: String)
+}
