--- composeApp/src/commonMain/kotlin/com/yong/taximeter/di/PlatformModule.kt
@@ -1,5 +1,0 @@
-package com.yong.taximeter.di
-
-import org.koin.core.module.Module
-
-expect val platformModule: Module
