+++ composeApp/src/iosMain/kotlin/com/yong/taximeter/MainViewController.kt
@@ -4,7 +4,6 @@ import androidx.compose.ui.window.ComposeUIViewController
 import com.revenuecat.purchases.kmp.Purchases
 import com.revenuecat.purchases.kmp.PurchasesConfiguration
 import com.yong.taximeter.di.appModule
-import com.yong.taximeter.di.platformModule
 import org.koin.core.context.startKoin
 import secrets.Secrets
 
@@ -15,7 +14,7 @@ fun MainViewController() = ComposeUIViewController(
         Purchases.configure(builder.build())
 
         startKoin {
-            modules(appModule, platformModule)
+            modules(appModule)
         }
     }
 ) { App() }
